[ui] fix build crash with precompiled expo-modules-core
