+++ packages/expo-ui/ios/ExpoUIModule.swift
@@ -1,7 +1,6 @@
 // Copyright 2025-present 650 Industries. All rights reserved.
 
 import ExpoModulesCore
-import ExpoModulesWorklets
 
 public final class ExpoUIModule: Module {
   public func definition() -> ModuleDefinition {
@@ -17,13 +16,7 @@ public final class ExpoUIModule: Module {
 
     // MARK: - Observable State
 
-    Class(WorkletCallback.self) {
-      Constructor { (worklet: Worklet) -> WorkletCallback in
-        let callback = WorkletCallback()
-        callback.worklet = worklet
-        return callback
-      }
-    }
+    makeWorkletCallbackClass()
 
     Class(ObservableState.self) {
       Constructor { (initial: [String: Any]) -> ObservableState in
